Make Selfie's non-suspend memo methods `@JvmStatic`.
